Refactor Portfolio to build GH cards from scratch
