+++ src/components/Card/_card.scss
@@ -1,0 +1,11 @@
+.card {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  background-color: aquamarine;
+  color: black;
+  width: 50%;
+  height: 100px;
+  margin: 20px;
+}
