+++ src/index.scss
@@ -5,6 +5,5 @@
 @import './components/NavMenu/_navmenu';
 @import './components/About/_about';
 @import './components/Portfolio/_portfolio';
-@import './components/GitHubCards/_ghcards';
-@import './components/RepoCard/_repocard';
+@import './components/Card/_card';
 @import './components/Contact/_contact';
